Move shared test logic for FullStory destination tests into root destination test
