+++ packages/destination-actions/src/destinations/fullstory/__tests__/fullstory.test.ts
@@ -1,7 +1,22 @@
 import nock from 'nock'
 import { createTestEvent, createTestIntegration } from '@segment/actions-core'
 import Definition from '../index'
-import { forEachRegion, userId, anonymousId, email, displayName } from './test-support'
+import type { Settings } from '../generated-types'
+
+const regionToBaseUrlMapping: Record<string, string> = {
+  north_america: 'https://api.fullstory.com',
+  europe: 'https://api.eu1.fullstory.com'
+}
+
+const apiKey = 'fake-api-key'
+export const userId = 'fake-user-id'
+export const anonymousId = 'fake-anonymous-id'
+export const email = 'fake+email@example.com'
+export const displayName = 'fake-display-name'
+
+export const forEachRegion = (callback: (settings: Settings, baseUrl: string) => void) => {
+  Object.keys(regionToBaseUrlMapping).forEach((region) => callback({ apiKey, region }, regionToBaseUrlMapping[region]))
+}
 
 const testDestination = createTestIntegration(Definition)
