--- packages/destination-actions/src/destinations/fullstory/__tests__/test-support.ts
@@ -1,16 +1,0 @@
-import type { Settings } from '../generated-types'
-
-const regionToBaseUrlMapping: Record<string, string> = {
-  north_america: 'https://api.fullstory.com',
-  europe: 'https://api.eu1.fullstory.com'
-}
-
-const apiKey = 'fake-api-key'
-export const userId = 'fake-user-id'
-export const anonymousId = 'fake-anonymous-id'
-export const email = 'fake+email@example.com'
-export const displayName = 'fake-display-name'
-
-export const forEachRegion = (callback: (settings: Settings, baseUrl: string) => void) => {
-  Object.keys(regionToBaseUrlMapping).forEach((region) => callback({ apiKey, region }, regionToBaseUrlMapping[region]))
-}
